refactor(runtime): 引入 runtime 层，core 彻底脱离 PySide6
- 拆分 core/state.py：纯配置类型(SenderConfig 等)移入 core/config/，
  VideoState 归入 core/models/，AppState 迁移至 runtime/app_state.py
- 新增 runtime/ 包：Runtime、ConfigManager、AccountManager、AppResources
- 统一启动编排：Runtime.bootstrap() 幂等，负责主题初始化、凭据加载、配置加载
- MainWindow 接收 Runtime 实例，不再自行创建 AppState 和管理初始化顺序
- 删除 utils/config_manager.py 和 utils/account_manager.py
- core/ 层零 PySide6 导入，可独立进行纯 Python 单元测试
