+++ src/danmaku_sender/core/config/api_auth_config.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ApiAuthConfig:
+    sessdata: str
+    bili_jct: str
+    use_system_proxy: bool
