+++ src/danmaku_sender/core/config/monitor_config.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel, ConfigDict, Field
+
+
+class MonitorConfig(BaseModel):
+    """监视器的配置数据"""
+    model_config = ConfigDict(validate_assignment=True)
+
+    refresh_interval: int = Field(default=60, ge=10)
+
+    # 复用全局设置
+    prevent_sleep: bool = True
+    use_system_proxy: bool = True
+
+    stats_baseline: float = Field(default=0.0, exclude=True)
