+++ src/danmaku_sender/runtime/__init__.py
@@ -1,0 +1,3 @@
+from .runtime import Runtime
+
+__all__ = ["Runtime"]
