+++ src/danmaku_sender/runtime/resources.py
@@ -1,0 +1,15 @@
+import logging
+
+from ..ui.theme_manager import ThemeManager
+
+
+class AppResources:
+    """
+    全局资源持有
+
+    统一管理基础设施资源。
+    """
+
+    def __init__(self) -> None:
+        self.logger: logging.Logger = logging.getLogger("App")
+        self.theme: ThemeManager = ThemeManager.instance()
