+++ src/danmaku_sender/ui/controllers/account_controller.py
@@ -6,9 +6,10 @@ from PySide6.QtCore import QObject, Signal
 from danmaku_sender.api.bili_api_client import BiliApiClient
 from danmaku_sender.core.models.account import AccountCredential
 from danmaku_sender.core.models.user import UserProfile
-from danmaku_sender.core.state import ApiAuthConfig, AppState
+from danmaku_sender.core.config import ApiAuthConfig
+from danmaku_sender.runtime.app_state import AppState
+from danmaku_sender.runtime.account_manager import AccountManager
 from danmaku_sender.ui.framework.concurrency import PoolTask
-from danmaku_sender.utils.account_manager import save_accounts
 
 
 logger = logging.getLogger("App.System.Account")
@@ -45,7 +46,7 @@ class AccountController(QObject):
         super().__init__(parent)
 
     @staticmethod
-    def save_credentials(state: AppState, profile: UserProfile | None):
+    def save_credentials(state: AppState, profile: UserProfile | None, account_manager: AccountManager):
         """同步 sessdata/bili_jct 到已保存账号列表并写盘"""
         if not state.sessdata or not state.bili_jct:
             return
@@ -62,7 +63,7 @@ class AccountController(QObject):
                 acc.name = profile.username
             state.saved_accounts.append(acc)
 
-        save_accounts(state.saved_accounts)
+        account_manager.save_accounts(state.saved_accounts)
         if state.saved_accounts:
             logger.info(f"已保存 {len(state.saved_accounts)} 个账号。")
