+++ src/danmaku_sender/ui/editor/page.py
@@ -15,7 +15,7 @@ from .dialogs import EditDanmakuDialog, TimeOffsetDialog, ArrayGeneratorDialog
 from danmaku_sender.ui.framework.style_loader import get_svg_icon
 from danmaku_sender.ui.controllers.editor_controller import EditorController
 from danmaku_sender.core.models.editor_types import EditorField, InsertPosition
-from danmaku_sender.core.state import AppState
+from danmaku_sender.runtime.app_state import AppState
 
 
 class EditorPage(QWidget):
